+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/mih-kopylov/versioner-maven-plugin/compare/3.1.0...HEAD)
+### Changed
+- CommitActor commits now not only `pom.xml` files, but all changed files
 
 ## [3.1.0](https://github.com/mih-kopylov/versioner-maven-plugin/compare/3.0.0...3.1.0) - 2019.10.05
 ### Added